feat(topic-list): use native iOS search
