+++ src/app/(tabs)/(messages)/_layout.tsx
@@ -1,5 +1,19 @@
 import { Stack } from 'expo-router';
+import { useThemeColor } from 'heroui-native/hooks';
+
+import { isIOS, isLiquidGlassAvailable } from '@/config/constants';
 
 export default function MessagesStackLayout() {
-  return <Stack screenOptions={{ headerShown: false }} />;
+  const foregroundColor = useThemeColor('foreground');
+
+  return (
+    <Stack
+      screenOptions={{
+        headerShadowVisible: isIOS ? undefined : false,
+        headerShown: isIOS,
+        headerTransparent: isLiquidGlassAvailable,
+        headerTintColor: foregroundColor,
+      }}
+    />
+  );
 }
